feat(sign up): basic sign-up feature
